webhook updater

+++ model/episode.go
@@ -10,7 +10,7 @@ import (
 type Episode struct {
 	ID        primitive.ObjectID `bson:"_id"`
 	ComicID   primitive.ObjectID `bson:"comic_id"`
-	No        float64            `bson:"no"`
+	No        int                `bson:"no"`
 	Name      string             `bson:"name"`
 	CreatedAt time.Time          `bson:"created_at,omitempty"`
 	UpdatedAt time.Time          `bson:"updated_at,omitempty"`
